Move `BasicChunk`, `getActiveChunksByUnit` into courses router
